nvim: fix hiding of build output
C compiles so fast that sometimes the compiler is finished before i3
gets a chance to show the build output window...This causes the hide
build command to be executed before the build output window is even
shown. Hiding the build output window a bit later sort of fixes it for
now, but this needs a more stable solution.