tmux: fix yanking to system clipboard
In some cases, the yanked selection would not end up completely in the
system clipboard. As it turns out, tmux nowadays has better support for
yanking to the system clipboard without the need to go through 3rd party
plugins, and on top of that this also fixes the problems with garbled
selections.