nvim: improve async build from Neovim
Fix that in case a build finishes very quickly, the build output window
is not hidden because of synchronization issues between two separate
asynchronous shell commands.

Showing the build output is now handled using one separate asynchronous
shell command, fixing the aforementioned issue. On top of this,
configuration on the Vim side is now simpler.