mpd: allow using a git-annex music directory
Since git-annex points to the underlying data using symlinks, we have to
make sure mpd does not index songs twice. What is needed on top of this
is a file `.mpdignore` in the `.git` directory of the annex repository,
containing '*' to ignore everything inside that directory. This ensures
only the symlinks themselves end up in the mpd database.